refactor(router): move recognizers into the rules folder
