+++ modules/angular2/src/router/rules/path_recognizer.ts
