+++ modules/angular2/src/router/rules/recognizer.ts
