+++ modules/angular2/src/router/rules/regex_recognizer.ts
