+++ modules/angular2/test/router/regex_recognizer_spec.ts
@@ -10,8 +10,8 @@ import {
   SpyObject
 } from 'angular2/testing_internal';
 
-import {GeneratedUrlSegment} from 'angular2/src/router/recognizers/recognizer';
-import {RegexRecognizer} from 'angular2/src/router/recognizers/regex_recognizer';
+import {GeneratedUrlSegment} from 'angular2/src/router/rules/recognizer';
+import {RegexRecognizer} from 'angular2/src/router/rules/regex_recognizer';
 import {parser, Url} from 'angular2/src/router/url_parser';
 
 export function main() {
